Move feature code to feature module
