+++ feature/src/main/AndroidManifest.xml
@@ -1,5 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="br.com.jonathanarodr.playmovie">
+    package="br.com.jonathanarodr.playmovie.feature">
+
+    <application>
+
+        <activity
+            android:name=".ui.view.MainActivity"
+            android:exported="true">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
+
+        <activity
+            android:name=".ui.view.DetailActivity"
+            android:parentActivityName=".ui.view.MainActivity">
+            <meta-data
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".ui.view.MainActivity" />
+        </activity>
+
+    </application>
 
 </manifest>
