+++ feature/src/main/res/drawable/ic_event.xml
