+++ feature/src/main/res/drawable/ic_favorite_off.xml
