+++ feature/src/main/res/drawable/ic_series.xml
