+++ feature/src/main/res/layout/activity_detail.xml
