+++ feature/src/main/res/layout/list_item_movie.xml
