+++ feature/src/main/res/values/strings.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="title_movies">Filmes</string>
+    <string name="title_series">Séries</string>
+    <string name="title_favorite">Favoritos</string>
+
+    <string name="movie_detail_overview">Overview</string>
+
+    <string name="generic_message_ops_try_again">Ops, algo deu errado tente novamente</string>
+</resources>
